Change log level type from int to byte.

+++ pct/logger.go
@@ -47,7 +47,7 @@ func (l *Logger) Fatal(entry ...interface{}) {
 	l.log(proto.LOG_CRITICAL, entry)
 }
 
-func (l *Logger) log(level int, entry []interface{}) {
+func (l *Logger) log(level byte, entry []interface{}) {
 	fullMsg := ""
 	for i, str := range entry {
 		if i > 0 {
